refactor(stage6): move JSON contracts out of route layer
